requirements/requirements-pytorch.txt
torch==2.0.1
pytorch-lightning==2.0.1
gluonts==0.12.7
scikit-learn==1.2.2
matplotlib==3.6.3
pandas==2.0.1
scipy==1.10.1
einops==0.6.1
datasets==2.12.0
pyarrow==12.0.0
orjson==3.8.12
hydra-core==1.3.2
transformers==4.28.1
curated-transformers==1.1.0
tensorboard==2.15.0